// src/generated/models/Afb_resultatscreeningsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_resultatscreeningsafb_naturedelentite = {
  0: 'Tiers',
  1: 'Dirigeant',
  747010001: 'UBO'
} as const;
export type Afb_resultatscreeningsafb_naturedelentite = keyof typeof Afb_resultatscreeningsafb_naturedelentite;
export const Afb_resultatscreeningsafb_resultatducontrole = {
  0: 'N_gatif',
  1: 'Matchfaible',
  2: 'ErreurAPI',
  747010001: 'Matchpositif'
} as const;
export type Afb_resultatscreeningsafb_resultatducontrole = keyof typeof Afb_resultatscreeningsafb_resultatducontrole;
export const Afb_resultatscreeningsafb_typedecontrole = {
  0: 'Initial',
  1: 'Batchnocturne',
  747010001: 'Manuel'
} as const;
export type Afb_resultatscreeningsafb_typedecontrole = keyof typeof Afb_resultatscreeningsafb_typedecontrole;
export const Afb_resultatscreeningsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_resultatscreeningsstatecode = keyof typeof Afb_resultatscreeningsstatecode;
export const Afb_resultatscreeningsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_resultatscreeningsstatuscode = keyof typeof Afb_resultatscreeningsstatuscode;

export interface Afb_resultatscreeningsBase {
  afb_datedexecution: string;
  afb_detaildumatch?: string;
  afb_identifiantducontrole: string;
  afb_identifiantentite: string;
  afb_listesinterrogees: string;
  afb_naturedelentite: Afb_resultatscreeningsafb_naturedelentite;
  "afb_nomdutiers@odata.bind": string;
  afb_resultatducontrole: Afb_resultatscreeningsafb_resultatducontrole;
  afb_resultatscreeningid: string;
  afb_scoredeconfiance?: number;
  afb_typedecontrole: Afb_resultatscreeningsafb_typedecontrole;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_resultatscreeningsstatecode;
  statuscode?: Afb_resultatscreeningsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_resultatscreenings extends Afb_resultatscreeningsBase {
  afb_naturedelentitename?: string;
  afb_nomdutiersname?: string;
  afb_resultatducontrolename?: string;
  afb_typedecontrolename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_nomdutiers?: object;
  _afb_nomdutiers_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}